Se actualiza BackEnd

+++ src/producto/index.js
@@ -5,6 +5,8 @@ import dao from './DAO.js'
 import express from 'express'
 import _ from 'underscore'
 import mensajes from '../mensajes/mensajes.js'
+import jwt from 'jsonwebtoken'
+import tk from '../login/token.js'
 
 const router = express.Router()
 
@@ -21,14 +23,13 @@ router.get('/', (req, res) => {
     else if (req.query.id_producto) {
         dao.obtenerProductoPorId(req.query.id_producto).then(lista => {
             resultado = lista
-            if(resultado.length>0)
-            { 
+            if (resultado.length > 0) {
                 res.send(resultado)
             }
-            else{
+            else {
                 res.send(mensajes.mensajeSinResultados())
             }
-           
+
         })
 
     }
@@ -36,11 +37,10 @@ router.get('/', (req, res) => {
     else if (req.query.id_marca) {
         dao.obtenerProductoPorIdMarca(req.query.id_marca).then(lista => {
             resultado = lista
-            if(resultado.length>0)
-            { 
+            if (resultado.length > 0) {
                 res.send(resultado)
             }
-            else{
+            else {
                 res.send(mensajes.mensajeSinResultados())
             }
         })
@@ -50,11 +50,10 @@ router.get('/', (req, res) => {
     else if (req.query.id_tipo) {
         dao.obtenerProductoPorIdTipo(req.query.id_tipo).then(lista => {
             resultado = lista
-            if(resultado.length>0)
-            { 
+            if (resultado.length > 0) {
                 res.send(resultado)
             }
-            else{
+            else {
                 res.send(mensajes.mensajeSinResultados())
             }
         })
@@ -64,11 +63,10 @@ router.get('/', (req, res) => {
     else if (req.query.modelo) {
         dao.obtenerProductoPorModelo(req.query.modelo).then(lista => {
             resultado = lista
-            if(resultado.length>0)
-            { 
+            if (resultado.length > 0) {
                 res.send(resultado)
             }
-            else{
+            else {
                 res.send(mensajes.mensajeSinResultados())
             }
         })
@@ -78,11 +76,10 @@ router.get('/', (req, res) => {
     else if (req.query.descripcion) {
         dao.obtenerProductoPorDescripcion(req.query.descripcion).then(lista => {
             resultado = lista
-            if(resultado.length>0)
-            { 
+            if (resultado.length > 0) {
                 res.send(resultado)
             }
-            else{
+            else {
                 res.send(mensajes.mensajeSinResultados())
             }
         })
@@ -96,32 +93,50 @@ router.get('/', (req, res) => {
 })
 
 // POST Producto
-router.post('/', (req, res) => {
-    let resultado = null
-    dao.agregarProducto(req.body).then(producto => {
-        resultado = producto
-        res.send(resultado)
+router.post('/', tk.verificarToken, (req, res) => {
+    jwt.verify(req.token, 'claveSecreta', (error, authData) => {
+        if (!error && authData.user[0].ADMINISTRADOR == 'S') {
+            /*    let resultado = null */
+            dao.agregarProducto(req.body).then(producto => {
+                /*   resultado = producto */
+                res.send(producto)
+            })
+        }
+        else {
+            res.send(mensajes.mensajeRutaNoAutorizada(authData))
+        }
     })
-
-
 })
 
 // PUT Producto
-router.put('/:id', (req, res) => {
-    dao.modificarProducto(req.params.id,req.body).then(producto => {
-        res.send(producto)
+router.put('/:id', tk.verificarToken, (req, res) => {
+    jwt.verify(req.token, 'claveSecreta', (error, authData) => {
+        if (!error && authData.user[0].ADMINISTRADOR == 'S') {
+            dao.modificarProducto(req.params.id, req.body).then(producto => {
+                res.send(producto)
+            })
+        }
+        else {
+            res.send(mensajes.mensajeRutaNoAutorizada(authData))
+        }
     })
-
-
 })
 
 // DELETE Producto
-router.delete('/:id', (req, res) => {
-    let resultado = null
-    dao.eliminarProductoById(req.params.id).then(producto => {
-        resultado = producto
-        res.send(producto)
+router.delete('/:id', tk.verificarToken, (req, res) => {
+    jwt.verify(req.token, 'claveSecreta', (error, authData) => {
+        if (!error && authData.user[0].ADMINISTRADOR == 'S') {
+            let resultado = null
+            dao.eliminarProductoById(req.params.id).then(producto => {
+                resultado = producto
+                res.send(producto)
+            })
+        }
+        else {
+            res.send(mensajes.mensajeRutaNoAutorizada(authData))
+        }
     })
 })
+
 
 export default router
